+++ lua/miriani/scripts/options.lua
@@ -40,6 +40,7 @@ local options = {
   follow_interrupt = {descr="Interrupt speech when following.", value="no", group="screen reader", type="bool"},
   follow_direction_sounds = {descr="Play direction sounds when following or being dragged.", value="no", group="general", type="bool"},
   friendly_combat = {descr="Gag friendly (non-praelor) sector combat messages.", value="no", group="gags", type="bool"},
+  gag_others_piloting = {descr="Gag messages when others pilot in the room.", value="no", group="gags", type="bool"},
   pa_interrupt = {descr="Interrupt speech for public address (PA) messages.", value="no", group="screen reader", type="bool"},
   praelor_interrupt = {descr="Interrupt speech when detecting insectoid activity.", value="no", group="screen reader", type="bool"},
   scan_interrupt = {descr="Interrupt speech for scan coordinates.", value="starships", group="screen reader", type="enum", options={"starships", "everything", "off"}},